# src/api.yml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Usage api
  description: Usage api service APIs
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /metering
paths:
...
  /v1/status/correlationId/{correlationId}:
    get:
      description: |
        get status using correlationId
      operationId: getStatusServiceByCorrelationId
      x-eov-operation-handler: status
      parameters:
        - in: path
          name: correlationId
          schema:
            type: string
          required: true
          description: correlationId for fetching status
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Detailed'
        - $ref: '#/components/parameters/AccountOrPrefix'
      responses:
        '200':
          description: all success status report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusRequestResponse'
        '207':
          description: multi status report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusRequestResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/status/requestId/{requestId}:
    get:
      description: |
        get status using requestId
      operationId: getStatusServiceByRequestId
      x-eov-operation-handler: status
      parameters:
        - in: path
          name: requestId
          schema:
            type: string
          required: true
          description: requestId for fetching status
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Detailed'
        - $ref: '#/components/parameters/AccountOrPrefix'
      responses:
        '200':
          description: all success status report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusRequestResponse'
        '207':
          description: multi status report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusRequestResponse'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/status/usageEventId/{usageEventId}:
    get:
      description: |
        get status using usageEventId
      operationId: getStatusServiceByUsageEventId
      x-eov-operation-handler: status
      parameters:
        - in: path
          name: usageEventId
          schema:
            type: string
          required: true
          description: usageEventId for fetching status
        - $ref: '#/components/parameters/AccountId'
        - $ref: '#/components/parameters/Detailed'
        - $ref: '#/components/parameters/AccountOrPrefix'
      responses:
        '200':
          description: all success status report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusByUsageEventId'
        '207':
          description: multi status report
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatusByUsageEventId'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /status/aborted/v1:
    get:
      description: |
        get all aborted statuses. Only superusers can call this API.
      operationId: getAbortedStatus
      x-eov-operation-handler: status
      parameters:
        - in: query
          name: eventIdPresent
          schema:
            type: string
          required: false
          description: If not provided returns all aborted statuses. If set to true will only return event statuses. When false returns only request statuses.
        - in: query
          name: tenant
          schema:
            type: string
          required: false
          description: when not provided targets default db. If provided fetches aborted status from given tenant db.
      responses:
        '200':
          description: all aborted status
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Status'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/tenant:
    post:
      description: create the Tenants
      operationId: createTenant
      x-eov-operation-handler: tenant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tenant'
      responses:
        '202':
          description: accepted
          content:
            application/json:
              schema:
                type: object
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      description: gets details of all tenants
      operationId: getAllTenants
      x-eov-operation-handler: tenant
      responses:
        '200':
          description: Gets all tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tenant'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/tenant/{accountOrPrefix}:
    delete:
      description: deletes tenant record with given id
      operationId: deleteTenant
      x-eov-operation-handler: tenant
      parameters:
        - in: path
          name: accountOrPrefix
          schema:
            type: string
          required: true
          description: accountOrPrefix of the record
      responses:
        '200':
          description: deletes tenant record
          content:
            application/json:
              schema:
                type: object
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Tenant:
      required:
        - accountOrPrefix
        - name
        - type
      properties:
        accountOrPrefix:
          type: string
        name:
          type: string
        type:
          type: string
          enum:
            - accountId
            - cosPrefix
          default: accountId
        rhmTest:
          type: boolean
          default: true
    StatusRequestResponse:
      required:
        - status
        - eventStatus
      properties:
        status:
          $ref: '#/components/schemas/Status'
        statusStep:
          items:
            $ref: '#/components/schemas/StatusStep'
        eventStatus:
          type: array
          items:
            type: object
            properties:
              eventId:
                type: string
              status:
                type: string
            required:
              - eventId
              - status
    StatusByUsageEventId:
      required:
        - status
        - usageEvent
      properties:
        status:
          type: object
        statusStep:
          type: array
          items:
            type: object
        usageEvent:
          type: object
          properties:
            usage:
              type: object
              properties:
                eventId:
                  type: string
              additionalProperties: true
              required:
                - eventId
            enrichment:
              type: object
            metrics:
              type: array
              items:
                type: object
...
    Error:
      required:
        - errorCode
        - message
      properties:
        errorCode:
          oneOf:
            - type: string
            - type: number
        message:
          oneOf:
            - type: string
            - type: number
        details:
          type: object
...
    Status:
      required:
        - requestId
        - requestType
        - inputFileName
        - startTime
        - replayAttempt
      properties:
        id:
          type: string
        correlationId:
          type: string
        requestId:
          type: string
        requestType:
          type: string
        replayAttempt:
          type: number
        inputFileName:
          type: string
        startTime:
          type: number
        endTime:
          type: number
        accountId:
          type: string
        accountOrPrefix:
          type: string
          nullable: true
        eventId:
          type: string
          nullable: true
        type:
          type: string
        userResponseReturned:
          type: boolean
        finalResult:
          type: string
        status:
          type: string
        errorCode:
          type: string
          nullable: true
        errorResponseMessage:
          type: string
          nullable: true
      additionalProperties: true
    StatusStep:
      required:
        - statusId
        - action
        - startTime
        - attempt
        - replayAttempt
        - state
      properties:
        id:
          type: string
        statusId:
          type: string
        action:
          type: string
        startTime:
          type: number
        endTime:
          type: number
        attempt:
          type: number
        replayAttempt:
          type: number
        state:
          type: string
        message:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
...
  parameters:
    AccountOrPrefix:
      in: query
      name: accountOrPrefix
      schema:
        type: string
      required: false
      description: accountId or COS prefix associated with the request. If left empty, will look in the default bucket and only for requests with a null accountId.
    AccountId:
      in: query
      name: accountId
      schema:
        type: string
      required: false
      description: accountId associated with the request.
    Detailed:
      in: query
      name: detailed
      schema:
        type: string
      required: false
      description: flag to return detailed response. Superusers only